crmforapp/requirements.txt
djangorestframework==3.11.2
Django==3.1.8
psycopg2-binary==2.8.4